requirements.txt
# ===========================================
# LLM Belief Elicitation - Python Dependencies
# ===========================================

# Core
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Machine Learning / Deep Learning
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Interpretability
captum>=0.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
scipy>=1.11.0
joblib>=1.3.0
tqdm>=4.65.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# ===================
# Optional: RAG Support
# ===================
# Uncomment these if you need RAG functionality
# langchain>=0.1.0
# langchain-openai>=0.0.5
# faiss-cpu>=1.7.0
# tiktoken>=0.5.0

# ===================
# Development (optional)
# ===================
# pytest>=7.0.0
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0
